# server/api/swagger.yaml
openapi: 3.0.3
info:
  title: robolaunch Industry Cloud Platform - OpenAPI 3.0
  description: |
    API for robolaunch ICP platform structure.
  termsOfService: https://robolaunch.io/terms-and-conditions
  contact:
    email: info@robolaunch.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
- url: https://robolaunch.swagger.io/api/v3
...
paths:
  /platform:
    get:
      tags:
      - Platform
      summary: Get all active/available versions of robolaunch ICP
      description: All active versions of the robolaunch ICP are listed
      operationId: getPlatform
      parameters:
      - name: url
        in: query
        description: URL of manifest
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: https://raw.githubusercontent.com/robolaunch/platform/main/samples/platform.yaml
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRobolaunch'
...
  /environments:
    get:
      tags:
      - Environments
      summary: Get all active/available environments in all of the versions of robolaunch
        ICP
      description: Domains and environment are listed of every active/available version
        of the robolaunch ICP
      operationId: getPlatformImages
      parameters:
      - name: url
        in: query
        description: URL of manifest
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: https://raw.githubusercontent.com/robolaunch/platform/main/samples/images.yaml
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePlatformImages'
  /environments/{version}:
    get:
      tags:
      - Environments
      summary: Get environments of a version of robolaunch ICP
      description: Domains and environment are listed of an active/available version
        of the robolaunch ICP
      operationId: getVersionedPlatformImages
      parameters:
      - name: url
        in: query
        description: URL of manifest
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: https://raw.githubusercontent.com/robolaunch/platform/main/samples/images.yaml
      - name: version
        in: path
        description: Platform version
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0.1.3-prerelease.1
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVersionedPlatformImages'
  /environments/{version}/{domain}:
    get:
      tags:
      - Environments
      summary: Get environments of a version of robolaunch ICP in a specific domain
      description: Environments are listed for an active/available version of the
        robolaunch ICP
      operationId: getEnvironmentsOfDomain
      parameters:
      - name: url
        in: query
        description: URL of manifest
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: https://raw.githubusercontent.com/robolaunch/platform/main/samples/images.yaml
      - name: version
        in: path
        description: Platform version
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0.1.3-prerelease.1
      - name: domain
        in: path
        description: Domain name
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: robotics
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDomain'
  /environments/{version}/{domain}/environment:
    get:
      tags:
      - Environments
      summary: Get environment from a specific version of robolaunch ICP in a specific
        domain
      description: Environment is returned for an active/available version of the
        robolaunch ICP in a specific domain
      operationId: getEnvironment
      parameters:
      - name: url
        in: query
        description: URL of manifest
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: https://raw.githubusercontent.com/robolaunch/platform/main/samples/images.yaml
      - name: version
        in: path
        description: Platform version
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: 0.1.3-prerelease.1
      - name: domain
        in: path
        description: Domain name
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: robotics
      - name: base
        in: query
        description: Base version
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 0.2.6-alpha.19
      - name: application
        in: query
        description: Application name
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: pyspark
      - name: appVersion
        in: query
        description: Application version
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 3.5.0
      - name: ubuntu
        in: query
        description: Ubuntu distribution
        required: true
        style: form
        explode: true
        schema:
          type: string
          default: jammy
          enum:
          - focal
          - jammy
      - name: desktop
        in: query
        description: Ubuntu desktop (if exists)
        required: true
        style: form
        explode: true
        schema:
          type: string
          default: xfce
          enum:
          - xfce
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvironment'
components:
  schemas:
    Robolaunch:
      required:
      - contact
      - maintainers
      - platform
      - versions
      type: object
      properties:
        platform:
          type: string
          example: robolaunch Industry Cloud Platform
        contact:
          type: object
          additionalProperties:
            type: string
            example: github.com/robolaunch
        maintainers:
          type: array
          items:
            pattern: "^@[a-z\\d](?:[a-z\\d]|-(?=[a-z\\d])){0,38}$"
            type: string
            example: '@tunahanertekin'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/VersionMeta'
      example:
        maintainers:
        - '@tunahanertekin'
        - '@tunahanertekin'
        versions:
        - date: 2024-02-05
          docs: https://docs.robolaunch.cloud
          version: 0.1.2-prerelease.10
        - date: 2024-02-05
          docs: https://docs.robolaunch.cloud
          version: 0.1.2-prerelease.10
        contact:
          key: github.com/robolaunch
        platform: robolaunch Industry Cloud Platform
...
    PlatformImages:
      type: object
      properties:
        versions:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/VersionedPlatformImages'
      example:
        versions:
          key:
            organization: robolaunchio
            domains:
              key:
                environments:
                - application:
                    name: plain
                    icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                    alias: Ubuntu 20.04
                    description: |
                      "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                    version: 0.1.0
                  base:
                    desktop: xfce
                    ubuntu: focal
                    version: 0.2.6-alpha.24
                - application:
                    name: plain
                    icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                    alias: Ubuntu 20.04
                    description: |
                      "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                    version: 0.1.0
                  base:
                    desktop: xfce
                    ubuntu: focal
                    version: 0.2.6-alpha.24
                alias: Robotics
                description: |
                  "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
            repository: devspace
...
    VersionedPlatformImages:
      type: object
      properties:
        organization:
          type: string
          example: robolaunchio
        repository:
          type: string
          example: devspace
        domains:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Domain'
      example:
        organization: robolaunchio
        domains:
          key:
            environments:
            - application:
                name: plain
                icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                alias: Ubuntu 20.04
                description: |
                  "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                version: 0.1.0
              base:
                desktop: xfce
                ubuntu: focal
                version: 0.2.6-alpha.24
            - application:
                name: plain
                icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                alias: Ubuntu 20.04
                description: |
                  "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                version: 0.1.0
              base:
                desktop: xfce
                ubuntu: focal
                version: 0.2.6-alpha.24
            alias: Robotics
            description: |
              "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
        repository: devspace
    VersionMeta:
      required:
      - date
      - docs
      - version
      type: object
      properties:
        date:
          pattern: "^\\d{4}\\-(0[1-9]|1[012])\\-(0[1-9]|[12][0-9]|3[01])$"
          type: string
          example: 2024-02-05
        version:
          pattern: "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\\
            d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\\
            +([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$"
          type: string
          example: 0.1.2-prerelease.10
        docs:
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\\
            b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
          type: string
          example: https://docs.robolaunch.cloud
      example:
        date: 2024-02-05
        docs: https://docs.robolaunch.cloud
        version: 0.1.2-prerelease.10
...
    Domain:
      type: object
      properties:
        alias:
          type: string
          example: Robotics
        description:
          type: string
          example: |
            "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
        environments:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
      example:
        environments:
        - application:
            name: plain
            icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
            alias: Ubuntu 20.04
            description: |
              "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
            version: 0.1.0
          base:
            desktop: xfce
            ubuntu: focal
            version: 0.2.6-alpha.24
        - application:
            name: plain
            icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
            alias: Ubuntu 20.04
            description: |
              "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
            version: 0.1.0
          base:
            desktop: xfce
            ubuntu: focal
            version: 0.2.6-alpha.24
        alias: Robotics
        description: |
          "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
    Environment:
      required:
      - application
      - base
      type: object
      properties:
        application:
          $ref: '#/components/schemas/Application'
        base:
          $ref: '#/components/schemas/Base'
      example:
        application:
          name: plain
          icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
          alias: Ubuntu 20.04
          description: |
            "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
          version: 0.1.0
        base:
          desktop: xfce
          ubuntu: focal
          version: 0.2.6-alpha.24
    Application:
      required:
      - alias
      - description
      - icon
      - name
      - version
      type: object
      properties:
        name:
          type: string
          example: plain
        version:
          type: string
          example: 0.1.0
        alias:
          type: string
          example: Ubuntu 20.04
        description:
          type: string
          example: |
            "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
        icon:
          type: string
          example: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
      example:
        name: plain
        icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
        alias: Ubuntu 20.04
        description: |
          "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
        version: 0.1.0
    Base:
      required:
      - desktop
      - ubuntu
      - version
      type: object
      properties:
        ubuntu:
          type: string
          example: focal
        desktop:
          type: string
          example: xfce
        version:
          pattern: "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\\
            d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\\
            +([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$"
          type: string
          example: 0.2.6-alpha.24
      example:
        desktop: xfce
        ubuntu: focal
        version: 0.2.6-alpha.24
    ResponseRobolaunch:
      required:
      - message
      - response
      - success
      type: object
      properties:
        success:
          type: boolean
          description: Success of the request.
        message:
          type: string
          description: Response message.
        response:
          $ref: '#/components/schemas/Robolaunch'
      example:
        success: true
        response:
          maintainers:
          - '@tunahanertekin'
          - '@tunahanertekin'
          versions:
          - date: 2024-02-05
            docs: https://docs.robolaunch.cloud
            version: 0.1.2-prerelease.10
          - date: 2024-02-05
            docs: https://docs.robolaunch.cloud
            version: 0.1.2-prerelease.10
          contact:
            key: github.com/robolaunch
          platform: robolaunch Industry Cloud Platform
        message: message
...
    ResponsePlatformImages:
      required:
      - message
      - response
      - success
      type: object
      properties:
        success:
          type: boolean
          description: Success of the request.
        message:
          type: string
          description: Response message.
        response:
          $ref: '#/components/schemas/PlatformImages'
      example:
        success: true
        response:
          versions:
            key:
              organization: robolaunchio
              domains:
                key:
                  environments:
                  - application:
                      name: plain
                      icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                      alias: Ubuntu 20.04
                      description: |
                        "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                      version: 0.1.0
                    base:
                      desktop: xfce
                      ubuntu: focal
                      version: 0.2.6-alpha.24
                  - application:
                      name: plain
                      icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                      alias: Ubuntu 20.04
                      description: |
                        "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                      version: 0.1.0
                    base:
                      desktop: xfce
                      ubuntu: focal
                      version: 0.2.6-alpha.24
                  alias: Robotics
                  description: |
                    "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
              repository: devspace
        message: message
    ResponseVersionedPlatformImages:
      required:
      - message
      - response
      - success
      type: object
      properties:
        success:
          type: boolean
          description: Success of the request.
        message:
          type: string
          description: Response message.
        response:
          $ref: '#/components/schemas/VersionedPlatformImages'
      example:
        success: true
        response:
          organization: robolaunchio
          domains:
            key:
              environments:
              - application:
                  name: plain
                  icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                  alias: Ubuntu 20.04
                  description: |
                    "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                  version: 0.1.0
                base:
                  desktop: xfce
                  ubuntu: focal
                  version: 0.2.6-alpha.24
              - application:
                  name: plain
                  icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
                  alias: Ubuntu 20.04
                  description: |
                    "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
                  version: 0.1.0
                base:
                  desktop: xfce
                  ubuntu: focal
                  version: 0.2.6-alpha.24
              alias: Robotics
              description: |
                "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
          repository: devspace
        message: message
    ResponseDomain:
      required:
      - message
      - response
      - success
      type: object
      properties:
        success:
          type: boolean
          description: Success of the request.
        message:
          type: string
          description: Response message.
        response:
          $ref: '#/components/schemas/Domain'
      example:
        success: true
        response:
          environments:
          - application:
              name: plain
              icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
              alias: Ubuntu 20.04
              description: |
                "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
              version: 0.1.0
            base:
              desktop: xfce
              ubuntu: focal
              version: 0.2.6-alpha.24
          - application:
              name: plain
              icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
              alias: Ubuntu 20.04
              description: |
                "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
              version: 0.1.0
            base:
              desktop: xfce
              ubuntu: focal
              version: 0.2.6-alpha.24
          alias: Robotics
          description: |
            "Embark on your journey in robotics with the dedicated environment on robolaunch. Whether you're exploring algorithms, building prototypes, or testing simulations, this workspace provides the tools and resources to bring your robotic projects to life. Dive into the world of robotics development and unleash your creativity in this specialized environment on robolaunch."
        message: message
    ResponseEnvironment:
      required:
      - message
      - response
      - success
      type: object
      properties:
        success:
          type: boolean
          description: Success of the request.
        message:
          type: string
          description: Response message.
        response:
          $ref: '#/components/schemas/Environment'
      example:
        success: true
        response:
          application:
            name: plain
            icon: https://raw.githubusercontent.com/robolaunch/trademark/main/applications/ubuntu-logo.png
            alias: Ubuntu 20.04
            description: |
              "This robolaunch environment, featuring Ubuntu 20.04 and CUDA support, empowers developers with versatility across programming languages. Seamlessly switch between languages in this efficient platform, offering GPU-accelerated computing for a wide array of programming tasks."
            version: 0.1.0
          base:
            desktop: xfce
            ubuntu: focal
            version: 0.2.6-alpha.24
        message: message