# server/pkg/models/model_base.go
/*
 * robolaunch Industry Cloud Platform - OpenAPI 3.0
 *
 * API for robolaunch ICP platform structure.
 *
 * API version: 1.0.11
 * Contact: info@robolaunch.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package models

type Base struct {
	Ubuntu string `json:"ubuntu"`

	Desktop string `json:"desktop"`

	Version string `json:"version"`
}